pyproject.toml
[project]
name = "django-github-actions-vuln-scan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "pytest-django>=4.11.1",
]
